# perf: call `module.enableCompileCache()`