### Description

Calls the new `module.enableCompileCache()` API (https://github.com/nodejs/node/pull/54501) which is available since Node.js 22.8.0.

```
$ hyperfine --warmup 3 'node ./node_modules/vite/bin/vite.js build' # without this PR
Benchmark 1: node ./node_modules/vite/bin/vite.js build
  Time (mean ± σ):     236.5 ms ±   1.4 ms    [User: 187.1 ms, System: 41.8 ms]
  Range (min … max):   235.3 ms … 239.5 ms    12 runs
 
$ hyperfine --warmup 3 'node ./node_modules/vite/bin/vite.js build' # with this PR
Benchmark 1: node ./node_modules/vite/bin/vite.js build
  Time (mean ± σ):     208.1 ms ±   3.6 ms    [User: 161.5 ms, System: 36.3 ms]
  Range (min … max):   204.1 ms … 216.8 ms    14 runs
```

<!-- What is this PR solving? Write a clear description or reference the issues it solves (e.g. `fixes #123`). What other alternatives have you explored? Are there any parts you think require more attention from reviewers? -->

<!----------------------------------------------------------------------
Before creating the pull request, please make sure you do the following:

- Read the Contributing Guidelines at https://github.com/vitejs/vite/blob/main/CONTRIBUTING.md.
- Check that there isn't already a PR that solves the problem the same way. If you find a duplicate, please help us reviewing it.
- Update the corresponding documentation if needed.
- Include relevant tests that fail without this PR but pass with it.

Thank you for contributing to Vite!
----------------------------------------------------------------------->